Refactor test suite for maintainability and strategic value
- Consolidate 14+ test files into 5 strategic files (64% reduction)
- Create behavior-focused tests over implementation details
- Add GKE + KubeRay integration test for end-to-end workflow validation
- Streamline conftest.py from 495 → 25 lines (95% reduction)
- Organize test utilities in tests/helpers/ package structure
- Focus on integration testing for better bug detection
- Maintain excellent test coverage while improving maintainability
- Optimize for fast-paced startup development environment

New structure:
- test_core_functionality.py: Tool registry, unified manager, system interfaces
- test_managers.py: Manager behavior patterns and cross-integration
- test_kubernetes.py: Kubernetes/KubeRay operations and workflows
- test_mcp_server.py: End-to-end integration tests (enhanced)
- helpers/: Reusable test utilities and fixtures

Removed redundant files:
- test_base_managers.py, test_core_*_manager.py
- test_mcp_tools.py, test_tool_extensions.py
- test_kuberay_manager.py, test_kubernetes_manager.py

All tests pass with improved execution speed and strategic focus.
