+++ pytest.ini
@@ -24,6 +24,7 @@ markers =
     core: marks tests for core functionality changes
     tools: marks tests for MCP tool functionality
     ray_manager: marks tests for Ray manager functionality
+    gke: marks tests requiring GKE cluster access (requires GOOGLE_APPLICATION_CREDENTIALS and GKE_TEST_CLUSTER)
 asyncio_mode = auto
 filterwarnings =
     ignore::DeprecationWarning
