+++ tests/helpers/__init__.py
@@ -1,0 +1,41 @@
+"""Test helpers package for Ray MCP tests."""
+
+# Import commonly used items for convenience
+from .e2e import (
+    start_ray_cluster,
+    stop_ray_cluster,
+    submit_and_wait_for_job,
+    verify_cluster_status,
+    wait_for_job_completion,
+)
+from .fixtures import e2e_ray_manager
+from .utils import (
+    E2EConfig,
+    TestScripts,
+    TempScriptManager,
+    call_tool,
+    get_text_content,
+    parse_tool_result,
+    run_ray_cleanup,
+    wait_for_ray_shutdown,
+)
+
+__all__ = [
+    # E2E utilities
+    "start_ray_cluster",
+    "stop_ray_cluster", 
+    "submit_and_wait_for_job",
+    "verify_cluster_status",
+    "wait_for_job_completion",
+    # Fixtures
+    "e2e_ray_manager",
+    # General utilities
+    "E2EConfig",
+    "TestScripts",
+    "TempScriptManager",
+    "call_tool",
+    "get_text_content",
+    "parse_tool_result",
+    "run_ray_cleanup",
+    "wait_for_ray_shutdown",
+]
